// src/app/lib/generated/models/GalleryItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GalleryItem
 */
export interface GalleryItem {
    /**
     * 
     * @type {string}
     * @memberof GalleryItem
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof GalleryItem
     */
    displayName: string;
    /**
     * 
     * @type {boolean}
     * @memberof GalleryItem
     */
    unlocked: boolean;
    /**
     * 
     * @type {number}
     * @memberof GalleryItem
     */
    price: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof GalleryItem
     */
    pictures: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof GalleryItem
     */
    imageCount: number;
}

/**
 * Check if a given object implements the GalleryItem interface.
 */
export function instanceOfGalleryItem(value: object): value is GalleryItem {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('unlocked' in value) || value['unlocked'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    if (!('pictures' in value) || value['pictures'] === undefined) return false;
    if (!('imageCount' in value) || value['imageCount'] === undefined) return false;
    return true;
}

export function GalleryItemFromJSON(json: any): GalleryItem {
    return GalleryItemFromJSONTyped(json, false);
}

export function GalleryItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): GalleryItem {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'displayName': json['display_name'],
        'unlocked': json['unlocked'],
        'price': json['price'],
        'pictures': json['pictures'],
        'imageCount': json['image_count'],
    };
}

export function GalleryItemToJSON(json: any): GalleryItem {
    return GalleryItemToJSONTyped(json, false);
}

export function GalleryItemToJSONTyped(value?: GalleryItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'display_name': value['displayName'],
        'unlocked': value['unlocked'],
        'price': value['price'],
        'pictures': value['pictures'],
        'image_count': value['imageCount'],
    };
}